requirements.txt
GitPython
matplotlib
numpy
pandas
scipy